Updated to Latest AZ Rev a7f52e901967e79ba735115fb2b5f6fb4f7a3ba7
Changed Human, Undead and Gnomes to use a Crossbow (because was impossible to give gun skill to that classes, for some reason core don't give to them even if i put them correctly on the database).

Repaired that Blood Elf and Draenei don't start with their racial with the new enabled classes.